backend/pyproject.toml
[project]
name = "codemind-backend"
version = "0.1.0"
description = "CodeMind FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.28.0",
  "redis>=5.2.0",
  "pydantic-settings>=2.7.0",
  "python-jose[cryptography]>=3.3.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
  "ruff>=0.8.4",
  "mypy>=1.14.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
